Added lab1 plotter  (#1)
* Added simple plotting function

Added a simple plotting function that creates 3 line plots, one for each min cutoff and colors the regions in between which specify the range of the various medals.

* Adjusted fonts and colors

Increased figure and font sizes for better readability. Also, added different colors to each line plots markers to make them pop out more from the background.

* Added trendlines to each tier cutoff

Added trendlines to each tiers' cutoff with a configurable degree for the polynomial fit, which is 2 by default resulting in a hyperbolic fit.