Hotfix gold score
* Added base files for lab 1

The base template for the jupyter notebook used in this lab has been created.
A few temporary functions have been added which outline way the notebook will be structured.

* Updated readme

Added a list of team members and experiments to the readme file.

* Added lab1 plotter  (#1)

* Added simple plotting function

Added a simple plotting function that creates 3 line plots, one for each min cutoff and colors the regions in between which specify the range of the various medals.

* Adjusted fonts and colors

Increased figure and font sizes for better readability. Also, added different colors to each line plots markers to make them pop out more from the background.

* Added trendlines to each tier cutoff

Added trendlines to each tiers' cutoff with a configurable degree for the polynomial fit, which is 2 by default resulting in a hyperbolic fit.

* Added Lab1-Crawler

* Implemented a function to get webpage for a given year

Implemented a method named connect_to_webpage. This method retrieves the statistics webpage for a given year and outputs the whole HTML file of the webpage.

* Implemented a function to extract stats from HTML.

Implemented a method named extract_stats_of_webpage. This method extracts and returns the least points for gold, silver, and bronze medals.

* Implemented a function to report the stats for a given year range

Implemented a method name fetch_scores_and_medals that reports the min gold, silver and bronze score for a range of years.

* Implemented a method to save the stats in a file.

Implemented a function called fetch_and_save. This method will calculate statistics and then save them in a file. The file format would be csv.

* Changed plot function signature

Changed the plot function to get file path containing scores as input

* Resolved crawler's conflict with develop

Added a new function to change between formats.

Co-authored-by: amsabour <amsabour79@gmail.com>
Co-authored-by: amsabour <33648067+amsabour@users.noreply.github.com>

* Fixed a bug regarding gold scores

Fixed a bug that caused the gold scores to be lower than the real value.

* Fixed gold score bug

Fixed the bug where gold score is off by one .

Co-authored-by: Aryo Lotfi <32937042+aryol@users.noreply.github.com>
Co-authored-by: Aryo Lotfi <aryolotfi77@gmail.com>