Lab3 decorator finished (#7)
* Added the base classes of the Decorator Task

Added the test and implementation classes for decorator pattern task. The test file was given and was modified a bit according to the JUnit version and packages structure. Also the intended classes got implemented to resolve all compile-time errors.

* Compeleted the implementation of classes of decoration pattern task

In this commit, we have completed the implementation of all of the classes regarding the decorator pattern task. In other words, we have resolved all of the runtime errors of the tests, and now all of tests pass.