Lab 3 completed (#9)
* Lab3 decorator finished (#7)

* Added the base classes of the Decorator Task

Added the test and implementation classes for decorator pattern task. The test file was given and was modified a bit according to the JUnit version and packages structure. Also the intended classes got implemented to resolve all compile-time errors.

* Compeleted the implementation of classes of decoration pattern task

In this commit, we have completed the implementation of all of the classes regarding the decorator pattern task. In other words, we have resolved all of the runtime errors of the tests, and now all of tests pass.

* Lab3 bridge finished (#8)

* Added tests for bridge pattern

* Added in power and multiplication implementations

* Refactored PowerOperation interface

By passing in the multOp as a parameter in the functions, the PowerOperation objects don't need to store their multOps themselves, and changing between differnet multOps is easier and more convenient.

Co-authored-by: Aryo Lotfi <32937042+aryol@users.noreply.github.com>