Lab 9 Complete (#26)
* Lab3 decorator finished (#7)

* Added the base classes of the Decorator Task

Added the test and implementation classes for decorator pattern task. The test file was given and was modified a bit according to the JUnit version and packages structure. Also the intended classes got implemented to resolve all compile-time errors.

* Compeleted the implementation of classes of decoration pattern task

In this commit, we have completed the implementation of all of the classes regarding the decorator pattern task. In other words, we have resolved all of the runtime errors of the tests, and now all of tests pass.

* Lab3 bridge finished (#8)

* Added tests for bridge pattern

* Added in power and multiplication implementations

* Refactored PowerOperation interface

By passing in the multOp as a parameter in the functions, the PowerOperation objects don't need to store their multOps themselves, and changing between differnet multOps is easier and more convenient.

* Lab4 profile service (#14)

* Added the initial version of profile service

The current version of profile service is a restful service capable of creating, updating, and reading entries. But it doesn't use token yet and also doesn't use database as well.

* Added token checking to profile service

* Fixed a few bugs

Co-authored-by: Aryo Lotfi <aryolotfi77@gmail.com>

* Lab4 auth service (#13)

* Auth microservice added.

* Fixed a few bugs

* Lab4 user management service (#11)

* Added User Management Service

* Fixed a few bugs

* Lab4 api gateway service (#10)

* Added API Gateway service

* Fixed a few bugs

* Added config file

* Added few tests

* Lab5 api gateway service (#16)

* Added ApiGateway service

* Added viewing all users for admin

* Lab5 user management service (#18)

* Merged account management and profile services into one

* Added a test file for easier debugging

* Added doc generation to code

* Lab5 auth service (#17)

* Copied Auth Service from Lab4

* Added documentation generation code

* Lab5 book service (#15)

* Completed the book service.

The CRUD and search of books has completed in a single service.

* Fixed little 404 bug

Co-authored-by: Aryo Lotfi <aryolotfi77@gmail.com>

* Lab6 Completed (#20)

* Created dockerfile for Book Service

* Added dockerfile for the rest of the services and also added previous files

* Updated flask_autodoc to flask_selfdoc

* Added requests to APIGateway requirements

* Added FLASK_RUN_PORT environmental variable to Dockerfiles

Co-authored-by: Aryo Lotfi <aryolotfi77@gmail.com>

* Lab7 Completed (#22)

* Added add feature in calculator by following Example.pdf

* Added dividing scenarios

* Implemented dividing in calculator

* Added power scenarios

* Implemented power in calculator / Added operation-based scenarios

* Lab 8 completed (#24)

* Added project base files

* Updated PersonTests to be more covering

* Added TransformServiceTest to improve coverage

By adding these tests, the coverage percentages of the TransformService and User classes were boosted significantly.

* Added UserServiceImpl tests to increase covered code percentage

These tests were designed to test the old findById method which has now been deprecated in favor of a leaner/meaner new implementation.

Co-authored-by: Aryo Lotfi <32937042+aryol@users.noreply.github.com>
Co-authored-by: Aryo Lotfi <aryolotfi77@gmail.com>